requirements.txt
pyqt5
colorama
pandas
python-rapidjson